+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Master
+
+
+## Release 1.5.1
 * bugfix in cucim resampling
 * add 6mm body model
 * multilabel files contain label names in extended header